add mvp pattern
