+++ app/src/main/java/com/example/myapplication/Model/MainModel.java
@@ -1,0 +1,15 @@
+package com.example.myapplication.Model;
+
+import com.example.myapplication.Presenter.Contract;
+
+public class MainModel {
+    Contract.Presenter presenter;
+
+    public MainModel(Contract.Presenter presenter) {
+        this.presenter = presenter;
+    }
+
+//    public saveData(int data) {
+//        //
+//    }
+}
