+++ app/src/main/java/com/example/myapplication/Model/Post.java
@@ -1,0 +1,7 @@
+package com.example.myapplication.Model;
+
+class Post {
+    String title;
+    boolean Status;
+    String uri;
+}
