+++ app/src/main/java/com/example/myapplication/Presenter/Contract.java
@@ -1,0 +1,12 @@
+package com.example.myapplication.Presenter;
+
+public interface Contract {
+    interface View {
+        void showResult(int answer);
+    }
+
+    interface Presenter {
+        void addNum(int num1, int num2);
+    }
+
+}
