+++ app/src/main/res/layout/fragment_third.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ThirdFragment">
+    tools:context=".View.ThirdFragment">
 
     <TextView
         android:id="@+id/textview_third"
